+++ playground/src/env.d.ts
@@ -1,3 +1,5 @@
+/// <reference types="@rsbuild/core/types" />
+
 declare module '*.css';
 declare module '*.base64' {
   const value: string;